CG_population_inflator {grattan}	R Documentation

Forecasting capital gains

Description

Forecasting capital gains

Usage

CG_population_inflator(x = 1, from_fy, to_fy, forecast.series = "mean",
  cg.series)

CG_inflator(x = 1, from_fy, to_fy, forecast.series = "mean")

Arguments

x
To be inflated.

from_fy, to_fy
Financial years designating the inflation period.

forecast.series
One of "mean", "lower", "upper". What estimator to use in forecasts. "lower" and "upper" give the lower and upper boundaries of the 95% prediction interval.

cg.series
(Not implemented.)

Value

For CG_population_inflator, the number of individuals estimated to incur capital gains in fy_year. For CG_inflator, an estimate of the nominal value of (total) capital gains in to_fy relative to the nominal value in from_fy.

[Package grattan version 1.6.0.0 Index]