Offset {grattan}	R Documentation

General offset in C++

Description

Calculate the offset given a threshold, a maximum offset, and a taper.

Arguments

x
A vector of incomes etc.

y
The maximum offset available; the offset when x is zero.

a
The maximum value of x at which the maximum offset is available.

m
The taper rate (the negative slope).

[Package grattan version 1.6.0.0 Index]