age_grouper {grattan}	R Documentation

Age grouper

Description

Age grouper

Usage

age_grouper(age, interval = 10, min_age = 25, max_age = 75,
  breaks = NULL, labels = NULL)

Arguments

age
A numeric age (in years).

interval
How big should the age range be. 25-34 means interval = 10.

min_age
What is the upper bound of the lowest bracket? (min_age = 25 means 'Under 25' will be the lowest bracket.)

max_age
What is the lower bound of the highest bracket? (max_age = 75 means '75+' will be the bracket.)

breaks
Specify breaks manually.

labels
Specify the labels manually.

Value

An ordered factor giving age ranges (separated by hyphens) as specified.

Examples

age_grouper(42)
age_grouper(42, interval = 5, min_age = 20, max_age = 60)

[Package grattan version 1.6.0.0 Index]