aus_pop_qtr_age {grattan}	R Documentation

Australian estimated resident population by age and date

Description

Australian estimated resident population by age and date

Usage

aus_pop_qtr_age(date = NULL, age = NULL, tbl = FALSE, roll = TRUE,
  roll.beyond = FALSE)

Arguments

date
A vector of dates. If NULL, values for all dates are returned in a table. The dates need not be quarters, provided roll != FALSE,

age
A vector of (integer) ages from 0 to 100 inclusive. If NULL, all ages are returned.

tbl
Should a table be returned? If FALSE, a vector is returned.

roll
Should a rolling join be performed?

roll.beyond
Should inputs be allowed to go beyond the limits of data (without a warning)? This is passed to data.table's join, so options other than TRUE and FALSE are available. See ?data.table.

Value

A data.table or vector with values of the estimated resident population.

Examples

aus_pop_qtr_age(date = as.Date("2016-01-01"), age = 42)

[Package grattan version 1.6.0.0 Index]