bto {grattan}	R Documentation

Beneficiary tax offset

Description

Beneficiary tax offset

Usage

bto(benefit_amount, fy.year = NULL, rate1 = 0.15,
  benefit_threshold = 6000, tax_threshold = 37000, rate2 = 0.15)

Arguments

benefit_amount
The amount of Tax Offsetable benefit received by the taxpayer during the income year.

fy.year
The income year. Not used by default.

rate1
The coefficient in Division 2, section 13(2) of the Income Tax Assessment (1936 Act) Regulation 2015 (the regulations).

benefit_threshold
The amount of benefits above which the offset applies.

tax_threshold
The threshold at the upper conclusion of the lowest marginal tax rate in the words of the section 13(3) of the regulations.

rate2
The second coefficient in section 13(3) of the regulations.

Value

The beneficiary tax offset.

WARNING

This function disagrees with the ATO online calculator.

[Package grattan version 1.6.0.0 Index]