Carer Payment

Carer payment is available to those who provide constant for a person who has a physical, intellectual, or psychiatric disability. Note that many of the arguments relate to the individual who receives the care (indicated by not starting with 'carer_'). Payment is made to the carer and not to the person receiving the care.

carer_payment(Date = NULL, fy.year = NULL,
  carer_fortnightly_income = 0,
  carer_annual_income = carer_fortnightly_income * 26,
  carer_has_partner = FALSE, carer_n_dependants = 0L,
  carer_partner_fortnightly_income = 0,
  carer_partner_annual_income = carer_partner_fortnightly_income * 26,
  carer_assets_value = 0, carer_is_home_owner = FALSE,
  carer_illness_separated_couple = FALSE, dclad_eligible = FALSE,
  high_adat = FALSE, living_at_home = TRUE,
  receiving_other_payment = FALSE,
  care_receiver_fortnightly_income = 0,
  care_receiver_annual_income = care_receiver_fortnightly_income * 26,
  care_receiver_asset_value = 0, partner_fortnightly_income = 0,
  partner_annual_income = partner_fortnightly_income * 26,
  partner_asset_value = 0, children_fortnightly_income = 0,
  children_annual_income = children_fortnightly_income * 26,
  children_asset_value = 0, parents_fortnightly_income = 0,
  parents_annual_income = parents_fortnightly_income * 26,
  parents_asset_value = 0)

Arguments

Date, fy.year
The financial year. Currently only 2015-16 is supported (the most recent survey of income and housing results).

carer_fortnightly_income, carer_annual_income
Carer's income for means-testing purposes. Provide one but not both.

carer_has_partner
(logical, default: FALSE) Does the carer have a partner?

carer_n_dependants
How many dependants does the carer have? Default is zero.

carer_partner_fortnightly_income, carer_partner_annual_income
The carer's partner's income.

carer_assets_value
Total value of carer's household assets.

carer_is_home_owner
(logical, default: FALSE) Does the carer own their own home?

carer_illness_separated_couple
Is the couple separated by illness? (Affects the assets test.)

dclad_eligible
Is the person receiving care a DCLAD (Disability Care Load Assessment) qualifying child as defined in http://guides.dss.gov.au/guide-social-security-law/1/1/q/17 ?

high_adat
Does the person receiving care have a high ADAT (Adult Disability Assessment Tool) score as defined in http://guides.dss.gov.au/guide-social-security-law/1/1/a/78 ?

living_at_home
Does the person receiving care live at home with their parents?

receiving_other_payment
Is the care receiver receiving other social security payments?

care_receiver_fortnightly_income
Care receiver's fortnightly income

care_receiver_annual_income
Care receiver's annual income

care_receiver_asset_value
Care receiver's asset value

partner_fortnightly_income
Care receiver's partner's fortnightly income

partner_annual_income
Care receiver's partner's annual income

partner_asset_value
Care receiver's partner's asset value

children_fortnightly_income
Care receiver's children's fortnightly income

children_annual_income
Care receiver's children's annual income

children_asset_value
Care receiver's children's asset value

parents_fortnightly_income
Care receiver's parents' fortnightly income

parents_annual_income
Care receiver's parents' annual income

parents_asset_value
Care receiver's parents' asset value