cpi_inflator_general_date {grattan}	R Documentation

CPI for general dates

Description

CPI for general dates

Usage

cpi_inflator_general_date(from_nominal_price = 1, from_date, to_date, ...)

Arguments

from_nominal_price
(numeric) the nominal prices to be converted to a real price

from_date
(character, date-like) the 'date' contemporaneous to from_nominal_price. The acceptable forms are 'YYYY', 'YYYY-YY' (financial year), 'YYYY-MM-DD', and 'YYYY-Q[1-4]' (quarters). Note a vector cannot contain a mixture of date forms.

to_date
(character, date-like) the date at which the real price is valued (where the nominal price equals the real price). Same forms as for from_date

...
other arguments passed to cpi_inflator_quarters

Value

A vector of real prices in to_date dollars.

[Package grattan version 1.6.0.0 Index]