differentially_uprate_wage {grattan}	R Documentation

Differential uprating

Description

Apply differential uprating to projections of the Sw_amt variable.

Usage

differentially_uprate_wage(wage = 1, from_fy, to_fy, ...)

Arguments

wage
A numeric vector to be uprated.

from_fy
The financial year contemporaneous to wage, which must be a financial year of an available sample file – in particular, not after 2015-16.

to_fy
The target of the uprating. Passed to wage_inflator.

...
Other arguments passed wage_inflator.

Details

See vignette("differential-uprating").

Value

The vector wage differentially uprated to to_fy.

Author(s)

Hugh Parsonage and William Young

Examples

ws <- c(20e3, 50e3, 100e3)
from <- "2013-14"
to <- "2016-17"
differentially_uprate_wage(ws, from, to)
differentially_uprate_wage(ws, from, to) / wage_inflator(ws, from, to)

# Use a wage series:
if (requireNamespace("taxstats", quietly = TRUE)) {
  library(data.table)
  library(taxstats)
  WageGrowth <- data.table(fy_year = c("2017-18", "2018-19"),
                           r = c(0.0, 0.1))
  Wage201314 <- sample_file_1314[["Sw_amt"]]
  
  data.table(Wage_201314 = Wage201314,
             Wage_201819 = 
               differentially_uprate_wage(Wage201314, 
                                          from_fy = "2013-14",
                                          to_fy = "2018-19",
                                          wage.series = WageGrowth))
}

[Package grattan version 1.6.0.0 Index]