generic_inflator {grattan}	R Documentation

Generic inflator

Description

Used to inflate variables in the sample file when there is no clear existing index. Note this is an unexported function: it is not available to the end-user.

Usage

generic_inflator(vars, h, fy.year.of.sample.file = "2012-13",
  nonzero = FALSE, estimator = "mean", pred_interval = 80)

Arguments

vars
A character vector of those variables within .sample_file for which forecasts are desired.

h
An integer, how many years ahead should the inflator be targeted.

fy.year.of.sample.file
A string representing the financial year of .sample_file.

nonzero
Should the forecast be taken on all values, or just nonzero values?

estimator
What forecast element should be used: the point estimate ("mean"), or the upper or lower endpoint of a prediction interval?

pred_interval
If estimator is upper or lower, what prediction interval are these the end points of?

Value

A data table of two columns: variable containing vars and inflator equal to the inflator to be applied to that variable to inflate it ahead h years.

[Package grattan version 1.6.0.0 Index]