gni {grattan}	R Documentation

Gross National Income, Australia

Description

Gross national income, at contemporaneous prices (called 'current prices' by the ABS).

Usage

gni_qtr(date, roll = "nearest")

gni_fy(fy_year)

Arguments

date
A Date vector or character coercible thereto.

roll
Passed to data.table when joining.

fy_year
Character vector of financial years.

Value

For gni_qtr, the quarterly GNI for the nearest quarter date. For gni_fy the sum over the quarters in the financial year provided. If fy_year would provide incomplete data (i.e. only sum three or fewer quarters), a warning is issued. Dates or fy_year outside the available data is neither a warning nor an error, but NA.

Source

Australian Bureau of Statistics, Catalogue 5206.0. Series A2304354T.

[Package grattan version 1.6.0.0 Index]