grattan-package {grattan}	R Documentation

The grattan package.

Description

Grattan package

Details

Tax modelling and other common tasks for Australian policy analysts, in support of the Grattan Institute, Melbourne. <https://grattan.edu.au>

Author(s)

hugh.parsonage+grattanpackage@grattan.edu.au

hugh.parsonage@gmail.com

See Also

Useful links:

https://github.com/HughParsonage/grattan

Report bugs at https://github.com/HughParsonage/grattan/issues

[Package grattan version 1.6.0.0 Index]