install_taxstats {grattan}	R Documentation

Install 'taxstats' files

Description

The taxstats packages provide the sample files as released by the ATO. These packages are used for testing, but are not available through CRAN as they are too large.

Usage

install_taxstats(...)

Arguments

...
Arguments passed to install.packages.

[Package grattan version 1.6.0.0 Index]