inverse_average_rate {grattan}	R Documentation

Inverse average tax rate

Description

Inverse average tax rate

Usage

inverse_average_rate(average_rate, ..., .max = 1e+08)

Arguments

average_rate
The average tax rate (\frac{tax}{income})

...
Parameters passed to income_tax.

.max
The maximum income to test before ending the search. (Used only to prevent infinite loops.)

Value

The minimum income at which the average tax rate exceeds average_rate.

Examples

inverse_average_rate(0.2, fy.year = "2014-15")

[Package grattan version 1.6.0.0 Index]