Inverse income tax functions

Inverse income tax functions

inverse_income(tax, fy.year = "2012-13", zero.tax.income = c("maximum",
  "zero", "uniform", numeric(1)), ...)

Arguments

tax
The tax payable.

fy.year
The relevant financial year.

zero.tax.income
A character vector, ("maximum", "zero", "uniform", numeric(1)) Given that many incomes map to zero taxes, the income_tax function is not invertible there. As a consequence, the inverse function's value must be specified for tax = 0. "maximum" returns the maximum integer income one can have with a zero tax liability; "zero" returns zero for any tax of zero; "uniform" provides a random integer from zero to the maximum income with a zero tax. The value can also be specified explicitly.

...
Other arguments passed to income_tax. If tax or fy.year are vectors, these should be named vectors.

Value

The approximate taxable income given the tax payable for the financial year. See Details.

Details

This function has an error of $2.