lf_inflator {grattan}	R Documentation

Labour force inflators

Description

Labour force inflators

Usage

lf_inflator_fy(labour_force = 1, from_fy = "2012-13", to_fy,
  useABSConnection = FALSE, allow.projection = TRUE, use.month = 1L,
  forecast.series = c("mean", "upper", "lower", "custom"),
  forecast.level = 95, lf.series = NULL)

lf_inflator(labour_force = 1, from_date = "2013-06-30", to_date,
  useABSConnection = FALSE)

Arguments

labour_force
A numeric vector.

from_fy
Financial year of labour_force.

to_fy
Financial year for which the labour force is predicted.

useABSConnection
Should the function connect with ABS.Stat via an SDMX connection? If FALSE (the default), a pre-prepared index table is used. This is much faster and more reliable (in terms of errors), though of course relies on the package maintainer to keep the tables up-to-date. The internal data was updated on 2018-05-21 to include data up to 2018-04-01.

allow.projection
Logical. Should projections be allowed?

use.month
An integer (corresponding to the output of data.table::month) representing the month of the series used for the inflation.

forecast.series
Whether to use the forecast mean, or the upper or lower boundaries of the prediction intervals.

forecast.level
The prediction interval to be used if forecast.series is upper or lower.

lf.series
If forecast.series = 'custom', a data.table with two variables, fy_year and r. The variable fy_year consists of all financial years between the last financial year in the (known) labour force series and to_fy inclusive. The variable r consists of rates of labour force growth assumed in each fy_year, which must be 1 in the first year (to connect with the original labour force series).

from_date
The date of labour_force.

to_date
Dates as a character vector.

Details

lf_inflator is used on dates. The underlying data series is available every month.

Value

The relative labour force between to_date and for_date or to_fy and from_fy, multiplied by labour_force.

Author(s)

Hugh Parsonage and Tim Cameron

Source

ABS Cat 6202.0 http://www.abs.gov.au/ausstats/abs@.nsf/mf/6202.0?OpenDocument.

Examples

lf_inflator_fy(labour_force = 1, from_fy = "2012-13", to_fy = "2013-14")

library(data.table)
# Custom 1% growth over 2017-18 -> 2018.19
lf_inflator_fy(from_fy = "2017-18",
               to_fy = "2018-19",
               forecast.series = "custom", 
               lf.series = data.table(fy_year = c("2017-18", "2018-19"),
                                      r = c(0, 0.01)))
## Not run: 
lf_inflator(labour_force = 1, from_date = "2013-06-30", to_date = "2014-06-30")

## End(Not run)

[Package grattan version 1.6.0.0 Index]