lito {grattan}	R Documentation

Low Income Tax Offset

Description

The Low Income Tax Offset (LITO) is a non-refundable tax offset to reduce ordinary personal income tax for low-income earners.

Usage

.lito(input)

lito(income, max_lito = 445, lito_taper = 0.015, min_bracket = 37000)

Arguments

input
A keyed data.table containing the financial year and the input of every observation for which the LITO should be calculated. The input must have the following structure. The structure will not be checked.

fy_year

The financial year the LITO parameters should be obtained. This must be the key of the data.table.

income

The Taxable Income of the individual.

ordering

An integer sequence from 1 to nrow(input) which will be the order of the output.

income
Income of taxpayer

max_lito
The maximum LITO available.

lito_taper
The amount by which LITO should be shaded out or reduced for every additional dollar of taxable income.

min_bracket
The income at which the lito_taper applies.

Value

For .lito, the a numeric vector equal to the offset for each income and each financial year in input. For lito, a numeric vector equal to the offset for each income given the LITO parameters.

[Package grattan version 1.6.0.0 Index]