medicare_levy {grattan}	R Documentation

Medicare levy

Description

The (actual) amount payable for the Medicare levy.

Usage

medicare_levy(income, fy.year = "2013-14", Spouse_income = 0,
  sapto.eligible = FALSE, sato = NULL, pto = NULL,
  family_status = "individual", n_dependants = 0, .checks = TRUE)

Arguments

income
The taxable income. A vector of numeric values.

fy.year
The financial year. A character vector satisfying is.fy.

Spouse_income
The spouse's adjusted income.

sapto.eligible
(logical) Is the taxpayer eligible for SAPTO? See Details.

sato
Is the taxpayer eligible for the Senior Australians Tax Offset?

pto
Is the taxpayer eligible for the Pensions Tax Offset?

family_status
What is the taxpayer's family status: family or individual?

n_dependants
Number of children dependant on the taxpayer.

.checks
Should checks of certain arguments be made? Provided to improve performance when checks are not necessary.

Details

The Seniors and Pensioners Tax Offset was formed in 2012-13 as an amalgam of the Senior Australians Tax Offset and the Pensions Tax Offset. Medicare rates before 2012-13 were different based on these offsets. For most taxpayers, eligibility would be based on whether your age is over the pension age (currently 65). If sato and pto are NULL, sapto.eligible stands for eligibility for the sato and not pto. If sato or pto are not NULL for such years, only sato is currently considered. Supplying pto independently is currently a warning.

Value

The Medicare levy payable for that taxpayer.

[Package grattan version 1.6.0.0 Index]