new_income_tax {grattan}	R Documentation

New income tax payable Income tax payable with new tax brackets, tax rates etc

Description

New income tax payable Income tax payable with new tax brackets, tax rates etc

Usage

new_income_tax(income, new_tax_tbl)

Arguments

income
A vector of taxable incomes.

new_tax_tbl
A data.table with columns lower_bracket and marginal_rate for the new brackets and marginal rates.

Value

The income according to the new parameters.

[Package grattan version 1.6.0.0 Index]