new_medicare_levy {grattan}	R Documentation

New medicare levy

Description

Use a different way to calculate medicare levy.

Usage

new_medicare_levy(parameter_table)

Arguments

parameter_table
A data.table containing

switches

The value in a row specifying which different medicare function is to apply.

lower_threshold

What is the lower medicare threshold, below which no medicare levy is applied, above which a tapering rate applies.

taper

What is the taper above lower_threshold.

rate

The medicare levy applicable above the medicare thresholds.

lower_up_for_each_child

How much the lower threshold should increase with each n_dependants.

lower_family_threshold

The threshold as applied to families (i.e. couples)

Value

A function similar to medicare_levy.

[Package grattan version 1.6.0.0 Index]