pminC {grattan}	R Documentation

Parallel maximum

Description

A faster pmin().

Arguments

x
A numeric vector.

a
A single numeric value.

Value

The parallel minimum of the input values. The 0 versions are shortcuts for a = 0.

Note

This function will always be faster than pmin(x, a) when a is a single value, but can be slower than pmin.int(x, a) when x is short. Use this function when comparing a numeric vector with a single value.

[Package grattan version 1.6.0.0 Index]