progressivity {grattan}	R Documentation

Compute the progressivity

Description

Compute the progressivity

Usage

progressivity(income, tax, measure = c("Reynolds-Smolensky", "Kakwani"))

Arguments

income
Pre-tax income.

tax
Tax paid.

measure
Currently, only "Reynolds-Smolensky" progressivity is calculated:

G_Y - G_Z

where G_Y is the Gini coefficient of income and G_X is the Gini coefficient of post-tax income.

Value

The progressivity measure. Positive for progressive tax systems, and higher the value the more progressive the system.

Examples

I <- c(10e3, 20e3, 50e3, 100e3, 150e3)
progressivity(I, 0.3 * I) # zero
progressivity(I, income_tax(I, "2017-18"))

[Package grattan version 1.6.0.0 Index]