project_to {grattan}	R Documentation

A function for simple projections of sample files

Description

A function for simple projections of sample files

Usage

project_to(sample_file, to_fy, fy.year.of.sample.file = NULL, ...)

Arguments

sample_file
A sample file, most likely the 2012-13 sample file. It is intended that to be the most recent.

to_fy
A string like "1066-67" representing the financial year for which forecasts of the sample file are desired.

fy.year.of.sample.file
The financial year of sample_file. See project for the default.

...
Other arguments passed to project.

Value

A sample file of the same number of rows as sample_file with inflated values (including WEIGHT).

[Package grattan version 1.6.0.0 Index]