Rent assistance

The rent assistance to each individual payable by financial year.

rent_assistance(fortnightly_rent = Inf, per = "fortnight",
  fy.year = NULL, Date = NULL, n_dependants = 0L,
  has_partner = FALSE, .prop_rent_paid_by_RA = 0.75, max_rate = NULL,
  min_rent = NULL, sharers_provision_applies = FALSE,
  is_homeowner = FALSE, lives_in_sharehouse = FALSE)

Arguments

fortnightly_rent
The fortnightly rent paid by each individual. By default, infinity, so the maximum rent assistance is returned by default, since rent assistance is capped at a maximum rate. Note the criteria for board and lodging which can be found at http://guides.dss.gov.au/guide-social-security-law/3/8/1/70

per
Specifies the timeframe in which payments will be made. Can either take value "fortnight" or "annual".

fy.year
(character) The financial year over which rent assistance is to be calculated. When left as NULL, defaults to the user's financial year, unless max_rate and min_rent are both set. If fy.year is set, the annual payment is provided.

Date
(Date vector or coercible to such) An alternative to fy.year. If both fy.year and Date are provided, fy.year is ignored, with a warning. If Date is used, the fortnightly rent assistance is provided.

n_dependants
(integer) Number of dependent children. By default, 0L, so no children.

has_partner
(logical) Is each individual married? By default, FALSE.

.prop_rent_paid_by_RA
The proportion of the rent above the minimum threshold paid by rent assistance. Since it so happens that this value is constant over the period, it is set here rather than being added to the internal table.

max_rate
If not NULL, a numeric vector indicating for each individual the maximum rent assistance payable.

min_rent
If not NULL, a numeric vector indicating for each individual the minimum fortnightly rent above which rent assistance is payable. max_rate and min_rent must not be used when fy.year is set.

sharers_provision_applies
(logical, default: FALSE) Does the sharers provision apply to the parent payment? The list of functions can be found in table 2 column 4 http://guides.dss.gov.au/guide-social-security-law/3/8/1/10

is_homeowner
(logical, default: FALSE) Does the individual own their own home?

lives_in_sharehouse
(logical, default: FALSE) Does the individual live in a sharehouse?

Value

If fy.year is used, the annual rent assistance payable for each individual; if Date is used, the fortnightly rent assistance payable. If the arguments cannot be recycled safely, the function errors.

Examples

# current annual rent assistance
rent_assistance()

#> `fy.year` not set, so defaulting to fy.year = "2019-20"

#> `per` not set; calculating fortnightly payment.

#> [1] 138.86

# current fortnightly payment
rent_assistance(Date = Sys.Date())

#> `per` not set; calculating fortnightly payment.

#> [1] 138.8

# zero since no rent
rent_assistance(0, Date = "2016-01-02")

#> `per` not set; calculating fortnightly payment.

#> [1] 0

# Rent assistance is payable at 75c for every dollar over min rent
rent_assistance(101, max_rate = 500, min_rent = 100)

#> `per` not set; calculating fortnightly payment.

#> [1] 0.75

rent_assistance(500, max_rate = 500, min_rent = 100)

#> `per` not set; calculating fortnightly payment.

#> [1] 300