residential_property_prices {grattan}	R Documentation

Residential property prices in Australia

Description

Residential property prices indexes for the capital cities of Australia, and a weighted average for the whole country.

Usage

residential_property_prices

Format

A data.table of three columns and 522 observations:

Date

Date of the index

City

Capital city (or Australia (weighted average))

Residential_property_price_index

An index (100 = 2011-12-01) measuring the price change in all residential dwellings.

Source

ABS Cat 6416.0. http://www.abs.gov.au/ausstats/abs@.nsf/mf/6416.0.

[Package grattan version 1.6.0.0 Index]