sapto {grattan}	R Documentation

Seniors and Pensioner Tax Offset

Description

Seniors and Pensioner Tax Offset

Usage

sapto(rebate_income, fy.year, fill = 0, sapto.eligible = TRUE,
  Spouse_income = 0, family_status = "single", .check = TRUE)

Arguments

rebate_income
The rebate income of the individual.

fy.year
The financial year in which sapto is to be calculated.

fill
If SAPTO was not applicable, what value should be used?

sapto.eligible
Is the individual eligible for sapto?

Spouse_income
Spouse income whose unutilized SAPTO may be added to the current taxpayer. Must match family_status; i.e. can only be nonzero when family_status != "single".

family_status
Family status of the individual.

.check
Run checks for consistency of values. For example, ensuring no single individuals have positive Spouse_income.

[Package grattan version 1.6.0.0 Index]