small_business_tax_offset {grattan}	R Documentation

Small Business Tax Offset

Description

Small Business Tax Offset

Usage

small_business_tax_offset(taxable_income, basic_income_tax_liability,
  .dots.ATO = NULL, aggregated_turnover = NULL,
  total_net_small_business_income = NULL, fy_year = NULL,
  tax_discount = NULL)

Arguments

taxable_income
Individual's assessable income.

basic_income_tax_liability
Tax liability (in dollars) according to the method in the box in s 4.10(3) of the Income Tax Assessment Act 1997 (Cth). In general, basic_income_tax_liability is the ordinary tax minus offsets. In particular, it does not include levies (such as the Medicare levy or the Temporary Budget Repair Levy). For example, in 2015-16, an individual with an assessable income of \$100,000 had a basic tax liability of approximately \$25,000.

.dots.ATO
A data.table of tax returns. If provided, it must contain the variables Total_PP_BE_amt, Total_PP_BI_amt, Total_NPP_BE_amt, Total_NPP_BI_amt. If both .dots.ATO and either aggregated_turnover or total_net_small_business_income are provided, .dots.ATO takes precedence, with a warning.

If .dots.ATO contains the variable Tot_net_small_business_inc, it is used instead of the income variables.

aggregated_turnover
A numeric vector the same length as taxable_income. Only used to determine whether or not the offset is applicable; that is, the offset only applies if aggregated turnover is less than \$2M.

Aggregated turnover of a taxpayer is the sum of the following:

the taxpayer's annual turnover for the income year,

the annual turnover of any entity connected with the taxpayer's, for that part of the income year that the entity is connected with the taxpayer's

the annual turnover of any entity that is an affiliate of the taxpayer, for that part of the income year that the entity is affiliated with the taxpayer's

When you calculate aggregated turnover for an income year, do not include either:

the annual turnover of other entities for any period of time that the entities are either not connected with the taxpayer or are not the taxpayer's affiliate, or

amounts resulting from any dealings between these entities for that part of the income year that the entity is connected or affiliated with the taxpayer.

https://www.ato.gov.au/Business/Research-and-development-tax-incentive/Claiming-the-tax-offset/Steps-to-claiming-the-tax-offset/Step-3—Calculate-your-aggregated-turnover/

total_net_small_business_income
Total net business income within the meaning of the Act. For most taxpayers, this is simply any net income from a business they own (or their share of net income from a business in which they have an interest). The only difference being in the calculation of the net business income of some minors (vide Division 6AA of Part III of the Act).

fy_year
The financial year for which the small business tax offset is to apply.

tax_discount
If you do not wish to use the legislated discount rate from a particular fy_year, you can specify it via tax_discount. If both are provided, tax_discount prevails, with a warning.

Source

Basic income tax method s4-10(3) http://classic.austlii.edu.au/au/legis/cth/consol_act/itaa1997240/s4.10.html. Explanatory memorandum http://parlinfo.aph.gov.au/parlInfo/download/legislation/ems/r5494_ems_0a26ca86-9c3f-4ffa-9b81-219ac09be454/upload_pdf/503041.pdf.

[Package grattan version 1.6.0.0 Index]