wage_inflator {grattan}	R Documentation

Inflation using the Wage Price Index.

Description

Predicts the inflation of hourly rates of pay, between two financial years.

Usage

wage_inflator(wage = 1, from_fy, to_fy, useABSConnection = FALSE,
  allow.projection = TRUE, forecast.series = c("mean", "upper", "lower",
  "custom"), forecast.level = 95, wage.series = NULL)

Arguments

wage
The amount to be inflated (1 by default).

from_fy
A character vector of the form "2012-13" representing the FY ending that the wage index is to be taken (i.e. Q4 in that year). FY year must be 1996-97 or later.

to_fy
The FY ending that the wage index is to be taken.

useABSConnection
Should the function connect with ABS.Stat via an SDMX connection? If FALSE (the default), a pre-prepared index table is used. This is much faster and more reliable (in terms of errors), though of course relies on the package maintainer to keep the tables up-to-date. The internal data was updated on 2018-05-21 to include data up to 2018-Q1.

allow.projection
If set to TRUE the forecast package is used to project forward, if required.

forecast.series
Whether to use the forecast mean, or the upper or lower boundaries of the prediction intervals. A fourth option custom allows manual forecasts to be set.

forecast.level
The prediction interval to be used if forecast.series is upper or lower.

wage.series
If forecast.series = 'custom', how future years should be inflated. The future wage series can be provided in two ways: (1) a single value, to be the assumed rate of wage inflation in years beyond the known series, or (2) a data.table with two variables, fy_year and r. If (2), the variable fy_year must be a vector of all financial years after the last financial year in the (known) wage series and the latest to_fy inclusive. The variable r consists of rates of wage growth assumed in each fy_year.

Value

The wage inflation between the two years.

Examples

# Wage inflation
wage_inflator(from_fy = "2013-14", to_fy = "2014-15")

# Custom wage inflation
wage_inflator(from_fy = "2016-17",
              to_fy = "2017-18",
              forecast.series = "custom",
              wage.series = 0.05)

[Package grattan version 1.6.0.0 Index]