Youth unemployment

Youth unemployment

youth_unemployment(income = 0, assets = 0, fy.year = NULL, Date = NULL,
  has_partner = FALSE, has_dependant = FALSE, age = 23,
  lives_at_home = FALSE, independent = TRUE, unemployed = FALSE)

Arguments

income
Numeric vector of fortnightly income for the income test.

assets
Numeric vector of the value of assets. By default, income and assets are both zero, thus returning the maximum benefit payable.

fy.year
A character vector of valid financial years between "2000-01" and "2020-21" specifying which financial year the allowance is to be calculated.

Date
(Date vector or coercible to such). An alternative to fy.year to specify the period over which the allowance is calculated.

has_partner
(logical, default: FALSE) Does the individual have a partner?

has_dependant
(logical, default: FALSE) Does the indvidiual have any dependant children?

age
Age (only determines whether the 16-17 age or 18 or over rates will apply).

lives_at_home
(logical, default: FALSE) Is the individual a dependant who lives at home?

independent
(logical, default: TRUE) Should the individual be considered independent.

unemployed
(logical, default: FALSE) Is the individual unemployed?

Value

The fortnightly unemployment benefit payable for each entry. The function is vectorized over its arguments, with any length-1 argument recycled. (Other vector recycling is not supported and will result in an error.)