Differential uprating

Apply differential uprating to projections of the Sw_amt variable.

differentially_uprate_wage(wage = 1, from_fy, to_fy, ...)

Arguments

wage
A numeric vector to be uprated.

from_fy
The financial year contemporaneous to wage, which must be a financial year of an available sample file -- in particular, not after 2015-16.

to_fy
The target of the uprating. Passed to wage_inflator.

...
Other arguments passed wage_inflator.

Value

The vector wage differentially uprated to to_fy.

Details

See vignette("differential-uprating").

Examples

ws <- c(20e3, 50e3, 100e3)
from <- "2013-14"
to <- "2016-17"
differentially_uprate_wage(ws, from, to)

#> [1]  21411.35  52749.51 106555.75

differentially_uprate_wage(ws, from, to) / wage_inflator(ws, from, to)

#> [1] 1.0061469 0.9915069 1.0014384

# Use a wage series:
if (requireNamespace("taxstats", quietly = TRUE)) {
  library(data.table)
  library(taxstats)
  WageGrowth <- data.table(fy_year = c("2017-18", "2018-19"),
                           r = c(0.0, 0.1))
  Wage201314 <- sample_file_1314[["Sw_amt"]]

  data.table(Wage_201314 = Wage201314,
             Wage_201819 =
               differentially_uprate_wage(Wage201314,
                                          from_fy = "2013-14",
                                          to_fy = "2018-19",
                                          wage.series = WageGrowth))
}

#>         Wage_201314 Wage_201819
#>      1:        4540    5146.604
#>      2:       58734   64272.094
#>      3:       39953   43641.082
#>      4:      112494  125106.759
#>      5:       94213  104310.655
#>     ---                        
#> 258770:       24491   27108.885
#> 258771:       47704   52041.705
#> 258772:       46690   50931.746
#> 258773:        5132    5817.595
#> 258774:       49508   54037.205