Family tax benefit

Family tax benefit

family_tax_benefit(.data = NULL, id_hh = NULL, id = NULL,
  age = NULL, income = NULL, in_secondary_school = NULL,
  single_parent = NULL, other_allowance_benefit_or_pension = NULL,
  maintenance_income = NULL, maintenance_children = NULL,
  income_test_ftbA_1_bound = 51027, income_test_ftbA_2_bound = 94316,
  income_test_ftbB_bound = 5402, taper_ftbA_1 = 0.2,
  taper_ftbA_2 = 0.3, taper_ftbB = 0.2, per = "year", copy = TRUE)

Arguments

.data
data.table input. Each row is an individual. Columns must be have the same names

id_hh
household id, used to group households to determine eligiblity and number of children

id
individual id

age
numeric: age of each id

income
numeric: income of each id

in_secondary_school
logical column: does id attend secondary school?

single_parent
logical column: is id (a parent) single?

other_allowance_benefit_or_pension
logical column: does the individual receive a pension, benefit, or labour market program payment such as Youth Allowance?

maintenance_income
numeric: the amount of maintenance income the individual receives for the care of a child/children from a previous relationship

maintenance_children
integer: the number of children in the care of id for whom id receives maintenance

income_test_ftbA_1_bound
Lower bound for which reduction in ftb A max payment occurs at rate taper_ftbA_1.

income_test_ftbA_2_bound
Lower bound for which reduction in ftb A base payment occurs at rate taper_ftbA_1.

income_test_ftbB_bound
Lower bound for which reduction in ftb B payment occurs at rate taper_ftbB.

taper_ftbA_1
The amount at which ftb A max payment is reduced for each dollar earned above income_test_ftbA_1_bound.

taper_ftbA_2
The amount at which ftb A base payment is reduced for each dollar earned above income_test_ftbA_2_bound.

taper_ftbB
The amount at which ftb B payment is reduced for each dollar earned above income_test_ftbB_bound.

per
How often the payment will be made. At present payments can only be annually.

copy
(logical, default: TRUE) Should a copy of .data be made before the calculation? If FALSE, intermediate values will be assigned by reference to .data (if not NULL).