Income tax payable as a function of SAPTO

Income tax payable as a function of SAPTO

income_tax_sapto(income, fy.year = NULL, age = 42,
  family_status = "individual", n_dependants = 0L,
  return.mode = c("numeric", "integer"), .dots.ATO = NULL,
  allow.forecasts = FALSE, sapto.eligible, medicare.sapto.eligible,
  new_sapto_tbl = NULL)

Arguments

income
The individual assessable income.

fy.year
The financial year in which the income was earned. Only tax years from 2000-01 to 2016-17 are available. If fy.year is not given, the current financial year is used by default.

age
The individual's age.

family_status
For Medicare and SAPTO purposes.

n_dependants
An integer for the number of children of the taxpayer (for the purposes of the Medicare levy).

return.mode
The mode (numeric or integer) of the returned vector.

.dots.ATO
A data.frame that contains additional information about the individual's circumstances, with columns the same as in the ATO sample files. If .dots.ATO is a data.table, I recommend you enclose it with copy().

allow.forecasts
should dates beyond 2016-17 be permitted? Currently, not permitted.

sapto.eligible
Specify explicitly the eligibility for SAPTO. If missing, defaults to ages over 65.

medicare.sapto.eligible
Specify explicitly the eligibility for SAPTO with respect to the Medicare levy for low-income earners. If missing, defaults to ages over 65.

new_sapto_tbl
If not NULL, supplied to new_sapto. Otherwise, fy.year is passed to sapto.

Details

Used to cost simple changes to SAPTO.