Model Rent Assistance

Model Rent Assistance

model_rent_assistance(sample_file, baseline_fy = NULL,
  baseline_Date = NULL, Per = "fortnight",
  .Prop_rent_paid_by_RA = NULL, Max_rate = NULL, Min_rent = NULL,
  calc_baseline_ra = TRUE, return. = c("sample_file", "new_ra",
  "sample_file.int"))

Arguments

sample_file
A sample file having the same variables as the data.frame in the example.

baseline_fy, baseline_Date
(character) The financial year/date over which the baseline rent assistance is to be calculated. Only one can be provided.

Per
Specifies the timeframe in which payments will be made. Can either take value "fortnight" or "annual".

.Prop_rent_paid_by_RA
The proportion of the rent above the minimum threshold paid by rent assistance.

Max_rate
If not NULL, a numeric vector indicating for each individual the maximum rent assistance payable.

Min_rent
If not NULL, a numeric vector indicating for each individual the minimum fortnightly rent above which rent assistance is payable. max_rate and min_rent

calc_baseline_ra
(logical, default: TRUE) Should the income tax in baseline_fy or baseline_Date be included as a column in the result?

return.
What should the function return? One of tax, sample_file, or sample_file.int. If tax, the tax payable under the settings; if sample_file, the sample_file, but with variables tax and possibly new_taxable_income; if sample_file.int, same as sample_file but new_tax is coerced to integer.

Examples

library(data.table)
sample <-
  CJ(rent = 1:500,
     n_dependants = 0:3,
     has_partner = 0:1 > 0,
     is_homeowner = 0:1 > 0,
     lives_in_sharehouse = 0:1 > 0)
model_rent_assistance(sample,
                      baseline_fy = "2018-19",
                      .Prop_rent_paid_by_RA = 0.75,
                      Max_rate = 500,
                      Min_rent = 100)

#>        rent n_dependants has_partner is_homeowner lives_in_sharehouse
#>     1:    1            0       FALSE        FALSE               FALSE
#>     2:    1            0       FALSE        FALSE                TRUE
#>     3:    1            0       FALSE         TRUE               FALSE
#>     4:    1            0       FALSE         TRUE                TRUE
#>     5:    1            0        TRUE        FALSE               FALSE
#>    ---                                                               
#> 15996:  500            3       FALSE         TRUE                TRUE
#> 15997:  500            3        TRUE        FALSE               FALSE
#> 15998:  500            3        TRUE        FALSE                TRUE
#> 15999:  500            3        TRUE         TRUE               FALSE
#> 16000:  500            3        TRUE         TRUE                TRUE
#>        baseline_ra new_ra
#>     1:        0.00      0
#>     2:        0.00      0
#>     3:        0.00      0
#>     4:        0.00      0
#>     5:        0.00      0
#>    ---                   
#> 15996:      180.43    300
#> 15997:      180.43    300
#> 15998:      180.43    300
#> 15999:      180.43    300
#> 16000:      180.43    300