Newstart allowance

Newstart allowance

newstart_allowance(fortnightly_income = 0, annual_income = 0,
  has_partner = FALSE, partner_pensioner = FALSE, n_dependants = 0,
  nine_months = FALSE, isjspceoalfofcoahodeoc = FALSE,
  principal_carer = FALSE, fortnightly_partner_income = 0,
  annual_partner_income = 0, age = 22, fy.year = "2015-16",
  assets_value = 0, homeowner = FALSE, lower = 102, upper = 252,
  taper_lower = 0.5, taper_upper = 0.6, taper_principal_carer = 0.4,
  per = c("year", "fortnight"))

Arguments

fortnightly_income
'Ordinary income' received fortnightly within the meaning of s. 1068-G1 of the Social Security Act 1991.

annual_income
'Ordinary income' received annually.

has_partner
Does the individual have a partner?

partner_pensioner
Does the partner receive a pension?

n_dependants
How many dependant children does the individual have?

nine_months
If the person is over 60 years old, have they been receiving payments for over 9 continuous months?

isjspceoalfofcoahodeoc
Is the recipient a single job seeker principal carer, either of large family or foster child/ren, or who is a home or distance educator of child/ren?

principal_carer
Is the individual the parent with most of the day-to-day care of child. Defined in https://www.humanservices.gov.au/individuals/enablers/principal-carer-rules-parenting-payment/41456.

fortnightly_partner_income
Partner's 'Ordinary income' received fortnightly.

annual_partner_income
Partner's Ordinary income' received annually.

age
The individual's age.

fy.year
Financial year. Default is "2015-16".

assets_value
Total value of household assets. Details can be found at https://www.humanservices.gov.au/individuals/enablers/assets/30621.

homeowner
Is the individual a homeowner?

lower
Lower bound for which reduction in payment occurs at rate taper_lower (taper_principal_carer for principal carers).

upper
Upper bound for which reduction in payment occurs at rate taper_lower. Lower bound for which reduction in payment occurs at rate taper_upper. Note that for principal carers there is no upper bound.

taper_lower
The amount at which the payment is reduced for each dollar earned between the lower and upper bounds for non-principal carers.

taper_upper
The amount at which the payment is reduced for each dollar earned above the upper bound for non-principal carers.

taper_principal_carer
The amount at which the payment is reduced for each dollar earned above the lower bound for principal carers.

per
Specifies the timeframe in which payments will be made. Can either take value "fortnight" or "annual".

Source

http://classic.austlii.edu.au/au/legis/cth/consol_act/ssa1991186/s1068.html