A function for simple projections of tables of Australian Taxation Office tax returns.

A function for simple projections of tables of Australian Taxation Office tax returns.

project(sample_file, h = 0L, fy.year.of.sample.file = NULL,
  WEIGHT = 50L, excl_vars, forecast.dots = list(estimator = "mean",
  pred_interval = 80), wage.series = NULL, lf.series = NULL,
  .recalculate.inflators = FALSE, .copyDT = TRUE,
  check_fy_sample_file = TRUE, differentially_uprate_Sw = TRUE)

Arguments

sample_file
A sample file, most likely the 2012-13 sample file. It is intended that to be the most recent.

h
An integer. How many years should the sample file be projected?

fy.year.of.sample.file
The financial year of sample_file. If NULL, the default, the number is inferred from the number of rows of sample_file to be one of 2012-13, 2013-14, or 2014-15.

WEIGHT
The sample weight for the sample file. (So a 2% file has WEIGHT = 50.)

excl_vars
A character vector of column names in sample_file that should not be inflated. Columns not present in the 2013-14 sample file are not inflated and nor are the columns Ind, Gender, age_range, Occ_code, Partner_status, Region, Lodgment_method, and PHI_Ind.

forecast.dots
A list containing parameters to be passed to generic_inflator.

wage.series
See wage_inflator. Note that the Sw_amt will uprated by differentially_uprate_wage.

lf.series
See lf_inflator_fy.

.recalculate.inflators
(logical, default: FALSE. Should generic_inflator() or CG_inflator be called to project the other variables? Adds time.

.copyDT
(logical, default: TRUE) Should a copy() of sample_file be made? If set to FALSE, will update sample_file.

check_fy_sample_file
(logical, default: TRUE) Should fy.year.of.sample.file be checked against sample_file? By default, TRUE, an error is raised if the base is not 2012-13, 2013-14, or 2014-15 and a warning is raised if the number of rows in sample_file is different to the known number of rows in the sample files.

differentially_uprate_Sw
(logical, default: TRUE) Should the salary and wage column (Sw_amt) be differentially uprating using (differentially_uprate_wage)?

Value

A sample file of the same number of rows as sample_file with inflated values (including WEIGHT).

Details

We recommend you use sample_file_1213 over sample_file_1314, unless you need the superannuation variables, as the latter suggests lower-than-recorded tax collections.

Superannuation variables are inflated by a fixed rate of 5% p.a.

Examples

# install.packages('taxstats', repos = 'https://hughparsonage.github.io/drat')
if (requireNamespace("taxstats", quietly = TRUE) &&
    requireNamespace("data.table", quietly = TRUE)) {
  library(taxstats)
  library(data.table)
  sample_file <- copy(sample_file_1314)
  sample_file_1617 <- project(sample_file,
                              h = 3L, # to "2016-17"
                              fy.year.of.sample.file = "2013-14")
}