Youth allowance

Youth allowance

youth_allowance(ordinary_income = 0, fy.year = date2fy(Sys.Date()),
  age = 18, eligible_if_over22 = FALSE, has_partner = FALSE,
  living_at_home = FALSE, n_dependants = 0L,
  isjspceoalfofcoahodeoc = FALSE, is_student = TRUE,
  per = "fortnight", taper1 = 0.5, taper2 = 0.6,
  FT_YA_student_lower = 437, FT_YA_student_upper = 524,
  FT_YA_jobseeker_lower = 143, FT_YA_jobseeker_upper = 250,
  excess_partner_income_mu = 0.6)

Arguments

ordinary_income
Fortnightly individual income. Defualt is zero.

fy.year
Financial year. Default is current financial year.

age
The individual's age. Default is 18 years.

eligible_if_over22
To be eligible for Youth Allowance while over 22, recipients must either commence full-time study or an Australian apprenticeship having been in receipt of an income support payment for at least 6 out of the last 9 months since turning 22, or study an approved course in English where English is not their first language.

has_partner
Does the individual have a partner?

living_at_home
Does the individual live at home with their parents?

n_dependants
How many dependant children does the individual have?

isjspceoalfofcoahodeoc
Is the recipient a single job seeker principal carer, either of large family or foster child/ren, or who is a home or distance educator of child/ren?

is_student
Is the individual a student? Note that apprenctices are considered students.

per
How often the payment will be made. Default is fortnightly. At present payments can only be fortnightly.

taper1
The amount at which the payment is reduced for each dollar earned between the lower and upper bounds.

taper2
The amount at which the payment is reduced for each dollar earned above the upper bound.

FT_YA_student_lower
Student and apprentice lower bound for which reduction in payment occurs at rate taper1

FT_YA_student_upper
Student and apprentice upper bound for which reduction in payment occurs at rate taper1. Student and apprentice lower bound for which reduction in payment occurs at rate taper2.

FT_YA_jobseeker_lower
Jobseeker lower bound for which reduction in payment occurs at rate taper1

FT_YA_jobseeker_upper
Jobseeker upper bound for which reduction in payment occurs at rate taper1. Student and apprentice lower bound for which reduction in payment occurs at rate taper2.

excess_partner_income_mu
The amount at which the payment is reduced for each dollar earned by the individual's partner.